various minor tweaks (#7)
* separate pathbuilders and trajectorybuilder


* add DualParameter interface to seal Time, Arclength, and Internal parameters; add RobotState class to reduce dual numbers that are exposed to users


* make converting RobotState to/from Pose2dDual easier


* make converting RobotState to/from Pose2dDual easier


* refactor: update DualParameter usage across multiple files


---------
